requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
scikit-learn==1.5.1
pandas==2.2.2
requests==2.32.3
numpy==2.1.1
fastapi[standard]==0.115.0